В RoomDaoImpl добавил метод поиска свободных комнат в интервале дат по критериям OrderEntity, необходимо добавить комнаты с разными параметрами и покрыть тестами данный метод отель с id=269 как основа для тестов

+++ hotel/trunk/src/main/java/ru/yandex/zhmyd/hotel/model/DisplayedOrder.java
@@ -71,7 +71,7 @@ public class DisplayedOrder {
         return  order.getRoomCategory();
     }
 
-    public Integer getPlaces(){
+    public Byte getPlaces(){
         return order.getPlaces();
     }
 }
